Joined compilation and execution
